Fix x86_64 build

+++ init/property_service.c
@@ -464,9 +464,9 @@ static void load_persistent_properties()
                     || (sb.st_uid != 0)
                     || (sb.st_gid != 0)
                     || (sb.st_nlink != 1)) {
-                ERROR("skipping insecure property file %s (uid=%u gid=%u nlink=%d mode=%o)\n",
+                ERROR("skipping insecure property file %s (uid=%u gid=%u nlink=%u mode=%o)\n",
                       entry->d_name, (unsigned int)sb.st_uid, (unsigned int)sb.st_gid,
-                      sb.st_nlink, sb.st_mode);
+                      (unsigned int)sb.st_nlink, sb.st_mode);
                 close(fd);
                 continue;
             }
